Move BackendMock to mocks.dart
