+++ app/test/frontend/handlers_custom_api_test.dart
@@ -16,6 +16,7 @@ import '../shared/handlers_test_utils.dart';
 import '../shared/utils.dart';
 
 import 'handlers_test_utils.dart';
+import 'mocks.dart';
 import 'utils.dart';
 
 void tScopedTest(String name, Future func()) {
